# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a simple Petstore server. It is based off of the Swagger Petstore example. See http://petstore.swagger.io for comparison.",
    "version": "1.0.0",
    "title": "Simple Petstore",
    "contact": {
      "email": "james.schubert@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/pets": {
      "post": {
        "tags": [
          "pet"
        ],
        "summary": "Add a new pet to the store",
        "description": "",
        "operationId": "addPet",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Pet to add to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Pet created"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "pet"
        ],
        "summary": "Update a pet",
        "description": "",
        "operationId": "updatePet",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Pet to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid id"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
...
    "/orders": {
      "get": {
        "tags": [
          "store"
        ],
        "summary": "Gets info about orders in the store",
        "description": "",
        "operationId": "getOrders",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "placed",
                "approved",
                "delivered"
              ],
              "default": "placed"
            },
            "collectionFormat": "csv"
          },
          {
            "in": "query",
            "name": "_page",
            "description": "Page number to return",
            "required": true,
            "type": "integer",
            "format": "int64",
            "minimum": "1",
            "default": "1"
          },
          {
            "in": "query",
            "name": "_limit",
            "description": "Page size to return",
            "required": false,
            "type": "integer",
            "format": "int64",
            "minimum": "1",
            "default": "5"
          }
        ],
        "responses": {
          "200": {
            "description": "Order information",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            },
            "headers": {
              "Link": {
                "type": "string",
                "description": "prev/first/next/last Link header info"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "petId": {
          "type": "integer",
          "format": "int64"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "shipDate": {
          "type": "string",
          "format": "date-time"
        },
        "status": {
          "type": "string",
          "description": "Order Status",
          "enum": [
            "placed",
            "approved",
            "delivered"
          ]
        },
        "complete": {
          "type": "boolean",
          "default": false
        }
      },
      "xml": {
        "name": "Order"
      }
    },
...
    "Tag": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tag"
      }
    },
    "Pet": {
      "type": "object",
      "required": [
        "name",
        "photoUrls"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "categoryId": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "doggie"
        },
        "photoUrls": {
          "type": "array",
          "xml": {
            "name": "photoUrl",
            "wrapped": true
          },
          "items": {
            "type": "string"
          }
        },
        "tags": {
          "type": "array",
          "xml": {
            "name": "tag",
            "wrapped": true
          },
          "items": {
            "$ref": "#/definitions/Tag"
          }
        },
        "status": {
          "type": "string",
          "description": "pet status in the store",
          "enum": [
            "available",
            "pending",
            "sold"
          ]
        }
      },
      "xml": {
        "name": "Pet"
      }
    }
  }
}